Handle manual delete of tf resources gracefully
When you create something in Terraform but delete it manually, Terraform
should gracefully handle it. If the API returns an error when the resource
doesn't exist, the read function should check to see if the resource is available
first. If the resource isn't available, the function should set the ID to an empty
string so Terraform "destroys" the resource in state